More 64 bit updates for use of MSWindows HANDLE/WHnd representation
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/gui/branches/gnustep_testplant_branch@40236 72102866-910b-0410-8b05-ffd578937521

+++ Headers/Additions/GNUstepGUI/GSDragView.h
@@ -127,7 +127,7 @@
 		    action: (NSDragOperation)action
 		  position: (NSPoint)eventLocation
 		 timestamp: (NSTimeInterval)time
-		  toWindow: (int)dWindowNumber;
+		  toWindow: (NSInteger)dWindowNumber;
 - (NSWindow*) windowAcceptingDnDunder: (NSPoint)mouseLocation
 			    windowRef: (int*)mouseWindowRef;
